User roles added and implemented in controller, authorization for different role added.

+++ WebAPI/Controllers/V1/PostsController.cs
@@ -1,9 +1,11 @@
 ﻿using Application.Dto;
 using Application.Interfaces;
+using Infrastructure.Identity;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.OData.Query;
 using Swashbuckle.AspNetCore.Annotations;
+using System.Security.Claims;
 using WebAPI.Filters;
 using WebAPI.Filters.Helpers;
 using WebAPI.Wrapper;
@@ -46,6 +48,7 @@ public class PostsController : ControllerBase
     }
 
     [SwaggerOperation(Summary = "Retrieves all posts")]
+    [Authorize(Roles = UserRoles.Admin)]
     [EnableQuery]
     [HttpGet("[action]")]
     public IQueryable<PostDto> GetAll()
@@ -67,25 +70,41 @@ public class PostsController : ControllerBase
     }
 
     [SwaggerOperation(Summary = "Create a new post")]
+    [Authorize(Roles = UserRoles.User)]
     [HttpPost]
     public async Task<IActionResult> Create(CreatePostDto newPost)
     {
-        var post = await _postService.AddNewPostAsync(newPost);
+        var post = await _postService.AddNewPostAsync(newPost, User.FindFirstValue(ClaimTypes.NameIdentifier));
         return Created($"api/posts/{post.Id}", new Response<PostDto>(post));
     }
 
     [SwaggerOperation(Summary = "Update an existing post")]
+    [Authorize(Roles = UserRoles.User)]
     [HttpPut]
     public async Task<IActionResult> Update(UpdatePostDto updatePost)
     {
+        var userOwnsPost = await _postService.UserOwnsPostAsync(updatePost.Id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+        if(!userOwnsPost)
+        {
+            return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post!" });
+        }
+
         await _postService.UpdatePostAsync(updatePost);
         return NoContent();
     }
 
     [SwaggerOperation(Summary = "Delete an existing post by a unique id")]
+    [Authorize(Roles = UserRoles.AdminOrUser)]
     [HttpDelete]
     public async Task<IActionResult> Delete(int id)
     {
+        var userOwnsPost = await _postService.UserOwnsPostAsync(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+        var isAdmin = User.FindFirstValue(ClaimTypes.Role).Contains(UserRoles.Admin);
+        if (!isAdmin && !userOwnsPost)
+        {
+            return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post!" });
+        }
+
         await _postService.DeletePostAsync(id);
         return NoContent();
     }
